+++ checkpoint/orbax/checkpoint/checkpoint_manager.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""CheckpointManager, a sync implementation of AbstractCheckpointManager."""
+"""A class providing functionalities for managing multiple checkpoints."""
 
 import asyncio
 import dataclasses